experiment with Merge Queues
Merge Queues will run the checks on the commit that's about to merge to
target branch, and will abort if a failure is detected. The difference
with auto-merges is that the checks are re-run just before merging,
while auto-merges behaves as if the target branch have not changed since
the checks was last run.